Add unknown_gnupg_key-boo1100427 needle bypass boo#1100427
Trying to test instal-only test with an iso rebuilt with mksusecd  tool.
The remaining problem of https://bugzilla.suse.com/show_bug.cgi?id=1100427
is that CHECKSUMS file has an unknown GnuPG signature
so define here the new needle and
accept the warning to be able to continue.
(refer to related commit in os-autoinst-distri-opensuse)

Signed-off-by: Michel Normand <normand@linux.vnet.ibm.com>